+++ lib/Applications.ps1
@@ -1,15 +1,12 @@
-@(
-    @('core', 'Test-ScoopDebugEnabled'),
-    @('Helpers', 'New-IssuePrompt'),
-    @('json', 'ConvertToPrettyJson'),
-    @('manifest', 'Resolve-ManifestInformation'),
-    @('Dependencies', 'Resolve-DependsProperty'),
-    @('Versions', 'Clear-InstalledVersion')
-) | ForEach-Object {
-    if (!([bool] (Get-Command $_[1] -ErrorAction 'Ignore'))) {
-        Write-Verbose "Import of lib '$($_[0])' initiated from '$PSCommandPath'"
-        . (Join-Path $PSScriptRoot "$($_[0]).ps1")
-    }
+if ($__importedApplications__ -eq $true) {
+    return
+} else {
+    Write-Verbose 'Importing Applications'
+}
+$__importedApplications__ = $false
+
+'core', 'Helpers', 'json', 'manifest', 'Dependencies', 'Versions' | ForEach-Object {
+    . (Join-Path $PSScriptRoot "${_}.ps1")
 }
 
 #region Application instalaltion info file
@@ -181,6 +178,7 @@ function app_status($app, $global) {
     $status.hold = ($install_info.hold -eq $true)
     $status.bucket = $install_info.bucket
     $status.removed = $false
+    $status.url = $install_info.url
 
     $todo = $app
     if ($install_info.bucket) {
@@ -197,6 +195,7 @@ function app_status($app, $global) {
 
     if ($manifest.version) { $status.latest_version = $manifest.version }
 
+    $status.manifest = $manifest
     $status.outdated = $false
     if ($status.version -and $status.latest_version) {
         $status.outdated = (Compare-Version -ReferenceVersion $status.version -DifferenceVersion $status.latest_version) -ne 0
@@ -322,3 +321,5 @@ function Test-ResolvedObjectIsInstalled {
         return $false
     }
 }
+
+$__importedApplications__ = $true