+++ lib/core.ps1
@@ -1,11 +1,12 @@
-@(
-    @('Helpers', 'New-IssuePrompt'),
-    @('Helpers', 'New-IssuePrompt')
-) | ForEach-Object {
-    if (!([bool] (Get-Command $_[1] -ErrorAction 'Ignore'))) {
-        Write-Verbose "Import of lib '$($_[0])' initiated from '$PSCommandPath'"
-        . (Join-Path $PSScriptRoot "$($_[0]).ps1")
-    }
+if ($__importedCore__ -eq $true) {
+    return
+} else {
+    Write-Verbose 'Importing core'
+}
+$__importedCore__ = $false
+
+'Helpers' | ForEach-Object {
+    . (Join-Path $PSScriptRoot "${_}.ps1")
 }
 
 # Such format is need to prevent automatic conversion of JSON date https://github.com/Ash258/Scoop-Core/issues/26
@@ -941,18 +942,23 @@ function handle_special_urls($url) {
 }
 #endregion TODO: Extract lib/Download.ps1
 
+# Returns default architecutre if none provided
+# Returns the first architecture, which is not same as the default one
 function Resolve-ArchitectureParameter {
     [CmdletBinding()]
     param([String[]] $Architecture)
 
     process {
         $arch = default_architecture
+        $Architecture = $Architecture | Where-Object { ![String]::IsNullOrEmpty($_) }
 
         foreach ($a in $Architecture) {
-            if ($null -eq $a) { continue }
-
             try {
-                $arch = ensure_architecture $a
+                $newArch = ensure_architecture $a
+                if ($arch -ne $newArch) {
+                    $arch = $newArch
+                    break
+                }
             } catch {
                 Write-UserMessage -Warning -Message "'$a' is not a valid architecture. Detecting default system architecture"
             }
@@ -1065,3 +1071,5 @@ $PSNativeCommandArgumentPassing = 'Legacy'
 # Setup proxy globally
 setup_proxy
 #endregion Main
+
+$__importedCore__ = $true